+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 -   `raise_operator_error()` function to raise an error about an operator in a Pythonic style.
 -   `Scene.switch()` instance method that allows users to switch to a scene without calling `Game.set_scene(scene_id)`
 -   `Game.draw` and `Game.update` function, both of which are overrideable, to give user-defineable functionality not reliant on scenes.
+-   `Group.contains()` method for checking whether a group or gameobject has already been added to it.
 
 ### Changed
 
@@ -45,6 +46,7 @@
 -   Switching scenes now only takes effect on the next frame.
 -   Reordered `Draw.clear` params to be more intuitive.
 -   Renderer is now not automatically cleared if no scene exists.
+-   Renamed `Vector.random_inside_unit_circle()` to `Vector.rand_unit_vector()`
 
 ### Removed
 
@@ -59,6 +61,7 @@
 -   Updated all clone functions to work with new stuff.
 -   Changing the camera zoom now affects all Sprites.
 -   Made vector operations more complete and raise errors properly.
+-   `Group.count()` not working properly.
 
 ## [v2.2.0] - June 12, 2022